backend/requirement.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
backports-abc==0.5
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.0
django-celery==3.2.2
django-cors-middleware==1.3.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-filter==0.2.2
djangorestframework-jwt==1.11.0
idna==2.7
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
msgpack==0.5.6
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.6.4
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.19.1
salt==2018.3.2
six==1.11.0
tornado==4.5.3
urllib3==1.23
uWSGI==2.0.17.1
vine==1.1.4
